backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pinecone
PyMuPDF
Pillow
easyocr
python-dotenv
pydantic
